Move the landing page outside the steps framework to {host}/pilot/{reference}
