+++ app/controllers/pilot_controller.rb
@@ -1,0 +1,16 @@
+class PilotController < ApplicationController
+  before_action :validate_reference
+
+  def show; end
+
+  private
+
+  def reference
+    params[:id].to_s
+  end
+
+  def validate_reference
+    Participant.valid_reference?(reference) ||
+      (raise "Participant reference not found: '#{reference}'")
+  end
+end
