+++ app/views/pilot/show.en.html.erb
